move ConfUtils to tachyon.conf package
