+++ core/src/main/java/tachyon/conf/ConfUtils.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package tachyon.util;
+package tachyon.conf;
 
 import java.io.IOException;
 import java.util.Properties;
@@ -24,7 +24,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import tachyon.Constants;
-import tachyon.conf.TachyonConf;
 
 /**
  * Utility class for {@link tachyon.conf.TachyonConf}
